Move disk package under fsutil
