+++ internal/freeze/freezer.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/ipfs/go-datastore"
 	logging "github.com/ipfs/go-log/v2"
-	"github.com/ipni/storetheindex/internal/disk"
+	"github.com/ipni/storetheindex/fsutil/disk"
 )
 
 var log = logging.Logger("indexer/freezer")
